requirements.txt
livekit-agents
livekit
livekit-agents[groq]
livekit-agents[deepgram]~=1.3
livekit-plugins-deepgram~=1.3
livekit-plugins-silero
deepgram-sdk
llama-index
llama-index-embeddings-huggingface
python-dotenv
openai
groq
fastapi
uvicorn
torch
livekit-plugins-openai
livekit-plugins-assemblyai
livekit-plugins-cartesia
livekit-plugins-turn-detector
livekit-plugins-noise-cancellation
